requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
sqlalchemy==2.0.28
sqlite-utils==3.36.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3  # For future web scraping
 
